Relocate algo to properties file || Refactoring
